renamed melbourne-2019 directory to event-info
