+++ sections/event-info/event-schedule/event-schedule-data.model.ts
