+++ sections/event-info/event-schedule/event-schedule.module.scss
