+++ sections/event-info/event-schedule/event-schedule.tsx
