+++ sections/event-info/judges/judge-group.model.ts
