+++ sections/event-info/judges/judges.module.scss
