+++ sections/event-info/judges/judges.tsx
