+++ sections/event-info/prizes/prize-data.model.ts
