+++ sections/event-info/prizes/prizes.module.scss
