+++ sections/event-info/prizes/prizes.tsx
