+++ sections/event-info/sponsors/sponsor-data.model.ts
