+++ sections/event-info/sponsors/sponsors.module.scss
