+++ sections/event-info/sponsors/sponsors.tsx
